requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
twilio==8.3.0
requests==2.31.0
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
Pillow==10.1.0
aiohttp==3.9.3
python-multipart==0.0.6